Minor: rename components for consistency
* Add documentation to BaseRouteTab component
