+++ assets/vue/views/user/sessions/SessionsCurrent.vue
